services/ai/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
crewai[tools]
pytest
python-dotenv